LayoutTests/platform/ios-sim-deprecated/storage/indexeddb/mozilla/open-database-null-name-expected.txt

CONSOLE MESSAGE: line 19: TypeError: undefined is not an object (evaluating 'request.onsuccess = openSuccess')
Test IndexedDB: should NOT throw when opening a database with a null name

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open(null);
FAIL indexedDB.open(null); threw exception TypeError: null is not an object (evaluating 'indexedDB.open')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'request.onsuccess = openSuccess')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
